sorgenti/g3w-admin-iam-bari/requirements.txt
cryptography==38.0.4
pyOpenSSL==22.0.0
mozilla-django-oidc==2.0.0